// frontend/src/api/generated/models/bookmark-position-json.ts
/* tslint:disable */
/* eslint-disable */
/**
 * linkweave-api API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0-SNAPSHOT
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { Placement } from './placement';
import {
    PlacementFromJSON,
    PlacementFromJSONTyped,
    PlacementToJSON,
    PlacementToJSONTyped,
} from './placement';

/**
 * 
 * @export
 * @interface BookmarkPositionJson
 */
export interface BookmarkPositionJson {
    /**
     * 
     * @type {string}
     * @memberof BookmarkPositionJson
     */
    anchorBookmarkId: string;
    /**
     * 
     * @type {Placement}
     * @memberof BookmarkPositionJson
     */
    placement: Placement;
}



/**
 * Check if a given object implements the BookmarkPositionJson interface.
 */
export function instanceOfBookmarkPositionJson(value: object): value is BookmarkPositionJson {
    if (!('anchorBookmarkId' in value) || value['anchorBookmarkId'] === undefined) return false;
    if (!('placement' in value) || value['placement'] === undefined) return false;
    return true;
}

export function BookmarkPositionJsonFromJSON(json: any): BookmarkPositionJson {
    return BookmarkPositionJsonFromJSONTyped(json, false);
}

export function BookmarkPositionJsonFromJSONTyped(json: any, ignoreDiscriminator: boolean): BookmarkPositionJson {
    if (json == null) {
        return json;
    }
    return {
        
        'anchorBookmarkId': json['anchorBookmarkId'],
        'placement': PlacementFromJSON(json['placement']),
    };
}

export function BookmarkPositionJsonToJSON(json: any): BookmarkPositionJson {
    return BookmarkPositionJsonToJSONTyped(json, false);
}

export function BookmarkPositionJsonToJSONTyped(value?: BookmarkPositionJson | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'anchorBookmarkId': value['anchorBookmarkId'],
        'placement': PlacementToJSON(value['placement']),
    };
}